Fix compilation error caused by ClusteredDistribution parameters
