+++ sql-plugin/src/main/301until330-all/scala/com/nvidia/spark/rapids/shims/v2/GpuRangePartitioning.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020-2021, NVIDIA CORPORATION.
+ * Copyright (c) 2022, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,34 +14,16 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids
+package com.nvidia.spark.rapids.shims.v2
 
-import com.nvidia.spark.rapids.shims.v2.ShimExpression
+import com.nvidia.spark.rapids.GpuRangePartitioningBase
 
 import org.apache.spark.sql.catalyst.expressions.SortOrder
 import org.apache.spark.sql.catalyst.plans.physical.{ClusteredDistribution, Distribution, OrderedDistribution}
-import org.apache.spark.sql.types.{DataType, IntegerType}
-import org.apache.spark.sql.vectorized.ColumnarBatch
 
-/**
- * A GPU accelerated `org.apache.spark.sql.catalyst.plans.physical.Partitioning` that partitions
- * sortable records by range into roughly equal ranges. The ranges are determined by sampling
- * the content of the RDD passed in.
- *
- * @note The actual number of partitions created might not be the same
- * as the `numPartitions` parameter, in the case where the number of sampled records is less than
- * the value of `partitions`.
- *
- * The GpuRangePartitioner is where all of the processing actually happens.
- */
-case class GpuRangePartitioning(
-    gpuOrdering: Seq[SortOrder],
-    numPartitions: Int) extends GpuExpression with ShimExpression with GpuPartitioning {
-
-  override def children: Seq[SortOrder] = gpuOrdering
-  override def nullable: Boolean = false
-  override def dataType: DataType = IntegerType
-
+case class GpuRangePartitioning(gpuOrdering: Seq[SortOrder], numPartitions: Int)
+  extends GpuRangePartitioningBase(gpuOrdering, numPartitions) {
+  
   override def satisfies0(required: Distribution): Boolean = {
     super.satisfies0(required) || {
       required match {
@@ -71,7 +53,4 @@ case class GpuRangePartitioning(
       }
     }
   }
-
-  override def columnarEval(batch: ColumnarBatch): Any =
-    throw new IllegalStateException("This cannot be executed")
 }
